% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contmle.R
\name{contmle}
\alias{contmle}
\title{contmle}
\usage{
contmle(
  dt,
  estimation = list(outcome = list(fit = c("sl", "sl", "hal", "km"), model = Surv(time,
    delta == 1) ~ A, changepoint = NULL), cens = list(fit = c("sl", "sl", "hal", "km"),
    model = Surv(time, delta == 0) ~ A, changepoint = NULL)),
  target = 1,
  hal.screening = FALSE,
  one.step = FALSE,
  deps.size = 0.1,
  no.small.steps = 500,
  push.criterion = FALSE,
  iterative = FALSE,
  simultaneous.ci = FALSE,
  treat.model = A ~ L1 + L2 + L3,
  cut.off.cens = 0.001,
  separate.cr = FALSE,
  weighted.norm = c(FALSE, "sigma", "Sigma"),
  treat.effect = c("1", "0", "ate", "stochastic"),
  pi.star.fun = function(L) L$L1 * 0.2,
  tau = c(1.2),
  use.observed.times = FALSE,
  length.times = length(tau),
  check.times.size = NULL,
  sl.method = 3,
  V = 5,
  lambda.cv = NULL,
  maxit = 1000,
  init.lambda.cvs = c(sapply(1:5, function(jjj) (9:1)/(10^jjj))),
  lambda.cvs = seq(1e-07, 0.01, length = 50),
  lambda.grid.size = 50,
  penalize.time = FALSE,
  cut.covars = 8,
  cut.time = 10,
  cut.time.A = 10,
  cut.L.A = 8,
  cut.L.interaction = 3,
  maxIter = 10,
  verbose = FALSE,
  verbose.sl = FALSE,
  check.sup = FALSE,
  output.km = FALSE,
  only.km = FALSE,
  only.cox.sl = FALSE,
  output.RR = FALSE,
  sl.change.points = (0:12)/10,
  sl.models = list(mod1 = list(Surv(time, delta == 1) ~ A + L1 + L2 + L3), mod2 =
    list(Surv(time, delta == 1) ~ A + L1.squared + L2 + L3), mod3 = list(Surv(time, delta
    == 1) ~ L2.squared + A + L1.squared + L2 + L3), mod4 = list(Surv(time, delta == 1) ~
    A + L1.squared), mod5 = list(Surv(time, delta == 1) ~ A * L1 + L2 + L3), mod6 =
    list(Surv(time, delta == 1) ~ A * L1.squared + L2 + L3))
)
}
\arguments{
\item{dt}{[data.table] comprised of (t) time-to-event, (J) outcome, (A) intervention, (L) covariates, (ID) id}

\item{estimation}{[list] of outcome models:}

\item{target}{[numeric: 1] . Which event(s) to target}

\item{hal.screening}{[logical: F] use poisson-based (T/F?) or cox-based hal (T/F?)}

\item{one.step}{[logical: F] use one-step tmle (T) or iterative TMLE (F). Defaults to TRUE if competing risks are present)}

\item{deps.size}{[numeric: 0.1]}

\item{no.small.steps}{[numeric: 500]}

\item{push.criterion}{[logical: F]}

\item{iterative}{[logical: F] iterative TMLE}

\item{simultaneous.ci}{[logical: F] report simultaneous (T) or separate (F) inference}

\item{treat.model}{[formula] treatment model}

\item{cut.off.cens}{[numeric: 1e-3] cut-off for truncation of censoring weights.}

\item{separate.cr}{[logical:F] DO NOT CHANGE ##\@Helene maybe we should take this argument out?}

\item{weighted.norm}{[logical/character] c(FALSE, "sigma", "Sigma") # might be good to uniformize this}

\item{treat.effect}{[character] c("1", "0", "ate", "stochastic")}

\item{tau}{[numeric]}

\item{sl.models}{}
}
\value{

}
\description{
contmle
}
