% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.categorical.R
\name{fit.categorical}
\alias{fit.categorical}
\title{Pseudo hazard regression model for ordinal categoricals.}
\usage{
fit.categorical(formula, data, ...)
}
\arguments{
\item{formula}{Object of class \code{"formula"} that describes the
model to be fitted. The right hand side should be a
categorical factor variable or be coercible to factor.}

\item{data}{Optional data.frame in which argument \code{formula} is evaluated.}

\item{...}{Additional arguments that are passed to the call of \code{glm}
which is used to estimate the model.}
}
\description{
Pseudo hazard regression model for ordinal categoricals.
}
\details{
Suppose ordinal outcome \code{Y} takes three values \eqn{a,b,c,...,z}.
The pseudo hazard regression model specifies: \deqn{logit{P(Y=y|Y\ge y,X)}=\beta_y*X}
where \eqn{X} is a vector of covariates and \eqn{\beta_y} a vector of regression coefficients.
Specifically, the predicted probabilities from this model are
\deqn{P(Y=a|X)=\expit(\beta_a * X)}
\deqn{P(Y=b|X)=(1-P(Y=a|X))* \expit(\beta_b * X)}
\deqn{P(Y=c|X)=(1-P(Y=a|X))*(1-P(Y=b|X))* \expit(\beta_b * X)}
and so on.
}
\examples{
library(survival)
data(pbc)
f <- fit.categorical(edema~age+sex,data=pbc)
predict(f,newdata=pbc[1:3,])
}
